Fix tankDriveVolts negation
Tank drive (which is used by Autonomous) was negating right side drive motor voltage, which cancelled out the existing inversion done elsewhere, leading to spinning instead of forward driving. -CXH

An extra call to setInverted is also being made that should be examined, though it is setting the same value as a previous value so it's not causing harm at this time.